+++ e2e/commands/export.e2e.1.ts
@@ -64,8 +64,8 @@ describe('bit export command', function() {
       createFile('bar', 'foo2');
       createFile('baz', 'foo1');
       createFile('baz', 'foo2');
-      helper.command.runCmd('bit add bar -m foo1.js');
-      helper.command.runCmd('bit add baz -m foo1.js');
+      helper.command.addComponentDir('bar', { m: 'foo1.js' });
+      helper.command.addComponent('baz', { m: 'foo1.js' });
       helper.command.tagAllComponents();
       helper.command.exportAllComponents();
     });
@@ -82,8 +82,8 @@ describe('bit export command', function() {
       helper.scopeHelper.setNewLocalAndRemoteScopes();
       createFile('bar', 'foo1');
       createFile('bar', 'foo2');
-      helper.command.addComponent('bit add bar/foo1.js', { i: 'bar/foo1' });
-      helper.command.addComponent('bit add bar/foo2.js', { i: 'bar/foo2' });
+      helper.command.addComponent('bar/foo1.js', { i: 'bar/foo1' });
+      helper.command.addComponent('bar/foo2.js', { i: 'bar/foo2' });
       helper.command.tagAllComponents();
       // DO NOT change the next line to `helper.command.exportAllComponents()`. the current form catches some wierd bugs
       helper.command.exportComponent('bar/foo1 bar/foo2');
@@ -295,7 +295,7 @@ describe('bit export command', function() {
       fs.copySync(sourcePngFile, destPngFile);
       const stats = fs.statSync(destPngFile);
       pngSize = stats.size;
-      helper.command.runCmd('bit add bar -m foo.js -i bar/foo');
+      helper.command.addComponentDir('bar', { m: 'foo.js', i: 'bar/foo' });
       helper.command.tagAllComponents();
       helper.command.exportAllComponents();
     });