+++ src/api/consumer/lib/link.ts
@@ -1,10 +1,9 @@
 import { loadConsumer, Consumer } from '../../../consumer';
 import { linkAllToNodeModules } from '../../../links';
-import { LinksResult } from '../../../links/node-modules-linker';
 
-export default (async function linkAction(): Promise<LinksResult[]> {
+export default async function linkAction(changeRelativeToModulePaths: boolean) {
   const consumer: Consumer = await loadConsumer();
-  const linkResults = await linkAllToNodeModules(consumer);
+  const linkResults = await linkAllToNodeModules(consumer, changeRelativeToModulePaths);
   await consumer.onDestroy();
   return linkResults;
-});
+}