+++ src/api/consumer/lib/tag.ts
@@ -29,6 +29,7 @@ export async function tagAction(args: {
   verbose?: boolean;
   ignoreUnresolvedDependencies?: boolean;
   ignoreNewestVersion: boolean;
+  allowRelativePaths: boolean;
   skipTests: boolean;
   skipAutoTag: boolean;
 }): Promise<TagResults> {
@@ -41,6 +42,7 @@ export async function tagAction(args: {
     verbose,
     ignoreUnresolvedDependencies,
     ignoreNewestVersion,
+    allowRelativePaths,
     skipTests,
     skipAutoTag
   } = args;
@@ -65,7 +67,8 @@ export async function tagAction(args: {
     ignoreUnresolvedDependencies,
     ignoreNewestVersion,
     skipTests,
-    skipAutoTag
+    skipAutoTag,
+    allowRelativePaths
   );
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   tagResults.newComponents = newComponents;
@@ -98,6 +101,7 @@ export async function tagAllAction(args: {
   verbose?: boolean;
   ignoreUnresolvedDependencies?: boolean;
   ignoreNewestVersion: boolean;
+  allowRelativePaths: boolean;
   skipTests: boolean;
   scope?: string;
   includeImported?: boolean;
@@ -112,6 +116,7 @@ export async function tagAllAction(args: {
     verbose,
     ignoreUnresolvedDependencies,
     ignoreNewestVersion,
+    allowRelativePaths,
     skipTests,
     scope,
     includeImported,
@@ -147,7 +152,8 @@ export async function tagAllAction(args: {
     ignoreUnresolvedDependencies,
     ignoreNewestVersion,
     skipTests,
-    skipAutoTag
+    skipAutoTag,
+    allowRelativePaths
   );
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   tagResults.warnings = warnings;