+++ src/extensions/bit/bit.manifest.ts
@@ -1,7 +1,7 @@
+import { ExtensionManifest } from '@teambit/harmony';
 import { WorkspaceExt } from '../workspace';
 import { ScopeExt } from '../scope';
 import provideBit from './bit.provider';
-import { Scripts } from '../scripts';
 import { ComposerExt } from '../composer';
 import { InstallExt } from '../install';
 
@@ -9,9 +9,10 @@ import CompileExt from '../compile/compile.manifest';
 import TestExt from '../test/test.manifest';
 import { GraphExt } from '../graph';
 import { CreateExt } from '../create';
+import { FlowsExt } from '../flows';
 
 export default {
   name: 'Bit',
-  dependencies: [WorkspaceExt, ScopeExt, Scripts, CompileExt, TestExt, ComposerExt, GraphExt, InstallExt, CreateExt],
+  dependencies: [WorkspaceExt, ScopeExt, FlowsExt, CompileExt, TestExt, ComposerExt, GraphExt, InstallExt, CreateExt],
   provider: provideBit
-};
+} as ExtensionManifest;