+++ src/extensions/flows/index.ts
@@ -1,0 +1,2 @@
+export { default as FlowsExt } from './flows.manifest';
+export { Flows } from './flows';