+++ src/extensions/workspace/workspace.ts
@@ -103,6 +103,11 @@ export default class Workspace implements ComponentHost {
     return this.transformLegacyComponents(consumerComponents);
   }
 
+  async loadCapsules(bitIds: string[]) {
+    // throw new Error("Method not implemented.");
+    const components = await this.load(bitIds);
+    return components.map(comp => comp.capsule);
+  }
   /**
    * fully load components, including dependency resolution and prepare them for runtime.
    * @todo: remove the string option, use only BitId
@@ -112,7 +117,10 @@ export default class Workspace implements ComponentHost {
     const components = await this.getMany(ids);
     const isolatedEnvironment = await this.isolateEnv.createNetworkFromConsumer(
       components.map(c => c.id.toString()),
-      this.consumer
+      this.consumer,
+      {
+        packageManager: 'npm'
+      }
     );
     const capsulesMap = isolatedEnvironment.capsules.reduce((accum, curr) => {
       accum[curr.id.toString()] = curr.value;
@@ -157,7 +165,10 @@ export default class Workspace implements ComponentHost {
     main?: string,
     override = false
   ): Promise<AddActionResults> {
-    const addComponent = new AddComponents({ consumer: this.consumer }, { componentPaths, id, main, override });
+    const addComponent = new AddComponents(
+      { consumer: this.consumer },
+      { componentPaths, id, main, override, allowFiles: false }
+    );
     const addResults = await addComponent.add();
     // @todo: the legacy commands have `consumer.onDestroy()` on command completion, it writes the
     //  .bitmap file. workspace needs a similar mechanism. once done, remove the next line.